Sync selected Tunes playlists to a flash drive
This Applescript will sync an iTunes playlist folder with a flash drive
or directory on your hard disk.  It renames the genre of the copied
song to the playlist name.  Only the copied song’s genre is changed.
The script will also convert any Apple Lossless files to flac as the
file is copied.

This application was written specifically for the Tesla Model S which
supports genres but has no playlist function.